--- vscode/walkthroughs/search.md
@@ -1,9 +1,0 @@
-## Natural Language Search
-
-<img src="https://storage.googleapis.com/sourcegraph-assets/blog/vs-code-onboarding-walkthrough-dec-2023-natural-language.gif">
-
-Cody builds a Search index of your local files to make it easier to find what you’re looking for. Use a natural language query like “password hashing” or "connection retries" to quickly find and open the files that match your search.
-
-**✨ Pro-tips for using Cody’s Natural Language Search**
-<br>• Increase the quality of your chat responses by enabling "Search Context (Beta)" in your [Cody Settings](command:cody.status-bar.interacted). This enables chat to use the natural language search index as a context source.
-<br>• You can search file and symbol names too, not just natural language queries.
